pyproject.toml
[project]
name = "job-hunt-manager"
version = "0.1.0"
description = "Full-stack job hunt management application"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "alembic>=1.13.0",
    "aiosqlite>=0.20.0",
    "pydantic-settings>=2.5.0",
    "python-dotenv>=1.0.0",
    "cryptography>=43.0.0",
    "playwright>=1.48.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
]

[tool.hatch.build.targets.wheel]
packages = ["backend"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
